Add incomplete hygienic expect module
This is a hygienic rewrite of (ice-9 expect) macros, to fix
problems with the original non-higienic defmacro versions.

The original ice-9 expect has support for procedures as test
expressions in its clauses, but it keeps evaluating such
expressions upon every char read from the input port.

Instead the test expression creating the matcher procedure should be
evaluated only once, and that procedure should be stored in a binding,
and only referenced in the expect clauses themselves.

To make such changes, it seemed better to migrate the entire module
to use proper proper hygienic macros instead of defmacros.