Capture parameters via lexical bindings or named parameters
Ensures sensible default expressions for bindings, when parameters
bindings are not explicitly defined in the lexical context of the
macro expansion. Otherwise captures the available bindings.

Internally uses named parameters.

Alternatively the parameters can be parameterized directly via the named
parameters. These named parameters are not exported currently
to preserve backwards compatibility.

If parameters are not defined, then default values are used.